refactor: replace login and register
