+++ src/app/[locale]/login/page.tsx
@@ -1,5 +1,5 @@
-import Login from '~/components/login/login';
 import type { Locale } from '~/i18n-config';
+import Login from '~/widgets/login/login';
 import { getDictionary } from '../dictionaries';
 
 export default async function PageLogin({ params }: { params: Promise<{ locale: Locale; slug?: string[] }> }) {
